review

+++ resources/views/users/post-managerment.blade.php
@@ -35,7 +35,7 @@
 								<h5>{{ \Carbon\Carbon::parse($post->updated_at)->format('d/m/Y g:i A')}}</h5>
 							</td>
 							<td class="cart_description">
-								<input type = "button" class = "btn btn-danger btn-sm" value = "Delete" id = "del_user" onclick = "del_post({{$post->id}})">
+								<button class = "btn btn-danger btn-sm" value = "Delete" id = "del_user" onclick = "del_post({{$post->id}})"><span class="glyphicon glyphicon-trash"></span></button>
 							</td>
 							
 						</tr>
@@ -64,7 +64,9 @@
             });
 			function del_post(id){
 				console.log(id);
-				$.ajax({
+				var result = confirm("Are you sure to delete this post?");
+				if(result){
+					$.ajax({
                     type: 'delete',
                     url: '{{  url('del_post') }}',
                     data: ({
@@ -74,7 +76,9 @@
                     success:function(data){
 						$('#post').html(data);
                     }
-                });
+                	});
+				}
+				
 			}
 			$.ajaxSetup({
  			 headers: {
