+++ resources/views/users/user-management.blade.php
@@ -40,7 +40,7 @@
 							</td>
                             <td>
                                 @if($us->is_admin==0)
-								 <input type = "button" class = "btn btn-danger btn-sm" value = "Delete" id = "del_user" onclick = "del_user({{$us->id}})">
+								 <button type = "button" class = "btn btn-danger btn-sm" value = "Delete" id = "del_user" onclick = "del_user({{$us->id}})"><span class="glyphicon glyphicon-trash"></span></button>
                            		 
                                 @endif
 							</td>
@@ -68,19 +68,22 @@
                 });
             });
 			function del_user(id){
-				console.log(id);
-				$.ajax({
-                    type: 'delete',
-                    url: '{{  url('del_user') }}',
-                    data: ({
-						_token : $('meta[name="csrf-token"]').attr('content'), 
-                        'id':id
-                    }),
-                    success:function(data){
-                        $('#user').html(data.users);
-						$('#post').html(data.posts);
-                    }
-                });
+               var result = confirm("Are you sure to delete this user?");
+                if(result)  {
+                    $.ajax({
+                        type: 'delete',
+                        url: '{{  url('del_user') }}',
+                        data: ({
+                            _token : $('meta[name="csrf-token"]').attr('content'), 
+                            'id':id
+                        }),
+                        success:function(data){
+                            $('#user').html(data.users);
+                            $('#post').html(data.posts);
+                        }
+                    });
+              }
+				
 			}
 			
 			$.ajaxSetup({
